[]  move indexer module to internal
